refactor(yahoo): split provider lib root
