+++ crates/providers/yahoo/src/constants.rs
@@ -1,0 +1,6 @@
+pub(crate) const PROVIDER_ID: &str = "yahoo_finance";
+pub(crate) const DEFAULT_USER_AGENT: &str = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) dynamo-rs/0.1 Safari/537.36";
+pub(crate) const QUOTE_PAGE_URL: &str = "https://finance.yahoo.com/quote/AAPL";
+pub(crate) const GET_CRUMB_URL: &str = "https://query1.finance.yahoo.com/v1/test/getcrumb";
+pub(crate) const QUOTE_SUMMARY_MODULES: &str =
+    "price,summaryDetail,defaultKeyStatistics,financialData,quoteType,summaryProfile";
